+++ src/Frontend/Projeto_git/Intellectify/src/components/BackgroundImage/BackgroundImage.css
