+++ src/Frontend/Projeto_git/Intellectify/src/pages/Cadastro/Cadastro.css
