+++ src/Frontend/Projeto_git/Intellectify/src/pages/Reset/Reset.css
