+++ src/Frontend/Projeto_git/Intellectify/src/pages/Reset_02/Reset_02.css
